+++ lib/modules/theme/theme_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: theme
-// Updated at: 2026-04-06 18:41:41
-// Commit: test(theme): implement widget integration tests for theme
+// Updated at: 2026-04-09 09:40:25
+// Commit: test(theme): add unit tests for theme business logic
 
 class themeComponentState {
-  final String commitRef = '162';
-  final String lastMessage = 'test(theme): implement widget integration tests for theme';
-  final DateTime timestamp = DateTime.parse('2026-04-06T18:41:41+05:00');
+  final String commitRef = '204';
+  final String lastMessage = 'test(theme): add unit tests for theme business logic';
+  final DateTime timestamp = DateTime.parse('2026-04-09T09:40:25+05:00');
 }